reference-skill-example/requirements.txt
Flask
Flask-CORS
requests
pyyaml